feat: split algorithms tests
