+++ public/app/config/config.js
@@ -4,8 +4,8 @@ const path = require('path');
 const isWindows = require('../utils/isWindows');
 const {
   ALGORITHM_TYPES,
-  AUTHOR_USER,
-  AUTHOR_GRAASP,
+  PROGRAMMING_LANGUAGES,
+  AUTHORS,
 } = require('../../shared/constants');
 
 // resolve path for windows '\'
@@ -28,105 +28,13 @@ const PRODUCT_NAME = 'Graasp Insights';
 const TMP_FOLDER = 'tmp';
 const DEFAULT_LANG = 'en';
 
-const PROGRAMMING_LANGUAGES = {
-  PYTHON: 'Python',
-};
-
-const GRAASP_ALGORITHMS = [
-  {
-    id: 'hash-users',
-    name: 'Hash users',
-    description:
-      "Hash the userId field from the 'actions' and 'appInstanceResources'",
-    filename: 'hash_users.py',
-    filepath: path.join(ALGORITHMS_FOLDER, 'hash_users.py'),
-    author: AUTHOR_GRAASP,
-    language: PROGRAMMING_LANGUAGES.PYTHON,
-    type: ALGORITHM_TYPES.ANONYMIZATION,
-  },
-  {
-    id: 'sanitize-users',
-    name: 'Sanitize users',
-    description:
-      'Scan the dataset for occurrences of user names and user IDs, and replace such occurrences with a hash of the corresponding user ID',
-    filename: 'sanitize_users.py',
-    filepath: path.join(ALGORITHMS_FOLDER, 'sanitize_users.py'),
-    author: AUTHOR_GRAASP,
-    language: PROGRAMMING_LANGUAGES.PYTHON,
-    type: ALGORITHM_TYPES.ANONYMIZATION,
-  },
-  {
-    id: 'suppress-geolocation',
-    name: 'Suppress geolocation',
-    description: "Suppress the 'geolocation' field from the 'actions'",
-    filename: 'suppress_geolocation.py',
-    filepath: path.join(ALGORITHMS_FOLDER, 'suppress_geolocation.py'),
-    author: AUTHOR_GRAASP,
-    language: PROGRAMMING_LANGUAGES.PYTHON,
-    type: ALGORITHM_TYPES.ANONYMIZATION,
-  },
-  {
-    id: 'suppress-data',
-    name: 'Suppress data',
-    description: "Suppress the 'data' field from the 'actions'",
-    filename: 'suppress_data.py',
-    filepath: path.join(ALGORITHMS_FOLDER, 'suppress_data.py'),
-    author: AUTHOR_GRAASP,
-    language: PROGRAMMING_LANGUAGES.PYTHON,
-    type: ALGORITHM_TYPES.ANONYMIZATION,
-  },
-  {
-    id: 'suppress-users',
-    name: 'Suppress users',
-    description: "Suppress the 'users' field from a dataset",
-    filename: 'suppress_users.py',
-    filepath: path.join(ALGORITHMS_FOLDER, 'suppress_users.py'),
-    author: AUTHOR_GRAASP,
-    language: PROGRAMMING_LANGUAGES.PYTHON,
-    type: ALGORITHM_TYPES.ANONYMIZATION,
-  },
-  {
-    id: 'suppress-app-instances-settings',
-    name: 'Suppress appInstances settings',
-    description: "Suppress the 'settings' field from the 'appInstances'",
-    filename: 'suppress_appInstances_settings.py',
-    filepath: path.join(ALGORITHMS_FOLDER, 'suppress_appInstances_settings.py'),
-    author: AUTHOR_GRAASP,
-    language: PROGRAMMING_LANGUAGES.PYTHON,
-    type: ALGORITHM_TYPES.ANONYMIZATION,
-  },
-  {
-    id: 'suppress-app-instances-resources-data',
-    name: 'Suppress appInstancesResources data',
-    description: "Suppress the 'data' field from the 'appInstanceResource'",
-    filename: 'suppress_appInstanceResources_data.py',
-    filepath: path.join(
-      ALGORITHMS_FOLDER,
-      'suppress_appInstanceResources_data.py',
-    ),
-    author: AUTHOR_GRAASP,
-    language: PROGRAMMING_LANGUAGES.PYTHON,
-    type: ALGORITHM_TYPES.ANONYMIZATION,
-  },
-  {
-    id: 'two-anonymize-geolocation',
-    name: '2-Anonymize geolocation',
-    description: `Ensure that for every combination of 'country', 'region', and 'city', there are at least two users containing that combination.
-      The corresponding fields are suppressed (from 'city' to 'country') when necessary`,
-    filename: 'two_anonymize_geolocations.py',
-    filepath: path.join(ALGORITHMS_FOLDER, 'two_anonymize_geolocations.py'),
-    author: AUTHOR_GRAASP,
-    language: PROGRAMMING_LANGUAGES.PYTHON,
-    type: ALGORITHM_TYPES.ANONYMIZATION,
-  },
-];
-
 const GRAASP_UTILS = {
   id: 'graasp-utils',
   filename: 'graasp_utils.py',
   filepath: path.join(ALGORITHMS_FOLDER, 'graasp_utils.py'),
   language: PROGRAMMING_LANGUAGES.PYTHON,
   type: ALGORITHM_TYPES.UTILS,
+  author: AUTHORS.GRAASP,
 };
 
 const USER_UTILS = {
@@ -135,7 +43,7 @@ const USER_UTILS = {
   filepath: path.join(ALGORITHMS_FOLDER, 'utils.py'),
   language: PROGRAMMING_LANGUAGES.PYTHON,
   type: ALGORITHM_TYPES.UTILS,
-  author: AUTHOR_USER,
+  author: AUTHORS.USER,
 };
 
 const ACCEPTED_PYTHON_VERSIONS = ['2.7', '3'];
@@ -152,13 +60,9 @@ module.exports = {
   PRODUCT_NAME,
   DATASETS_FOLDER,
   ALGORITHMS_FOLDER,
-  GRAASP_ALGORITHMS,
   ALGORITHMS_FOLDER_NAME,
   escapeEscapeCharacter,
   ACCEPTED_PYTHON_VERSIONS,
-  PROGRAMMING_LANGUAGES,
-  AUTHOR_GRAASP,
-  AUTHOR_USER,
   GRAASP_UTILS,
   USER_UTILS,
 };
