+++ public/shared/constants.js
@@ -8,17 +8,20 @@ const ALGORITHM_TYPES = {
 };
 
 const APP_BACKGROUND_COLOR = '#F7F7F7';
+
 const PROGRAMMING_LANGUAGES = {
   PYTHON: 'Python',
+};
+
+const AUTHORS = {
+  GRAASP: 'Graasp',
+  USER: 'User',
 };
 
 const SCHEMA_TYPES = {
   GRAASP: 'GRAASP',
   NONE: 'NONE',
 };
-
-const AUTHOR_GRAASP = 'Graasp';
-const AUTHOR_USER = 'User';
 
 const EXECUTION_STATUSES = {
   PENDING: 'pending',
@@ -37,11 +40,10 @@ const EXECUTION_NAME_MAX_LENGTH = 25;
 module.exports = {
   ALGORITHM_TYPES,
   APP_BACKGROUND_COLOR,
+  AUTHORS,
   SCHEMA_TYPES,
   PROGRAMMING_LANGUAGES,
   EXECUTION_STATUSES,
   DATASET_TYPES,
-  AUTHOR_GRAASP,
-  AUTHOR_USER,
   EXECUTION_NAME_MAX_LENGTH,
 };
