+++ test/utils.js
@@ -24,3 +24,9 @@ export const clearInput = async (elem) => {
   const backSpaces = new Array(value.length).fill('Backspace');
   await elem.setValue(backSpaces);
 };
+
+export const menuGoTo = async (client, selector) => {
+  await openDrawer(client);
+  const el = await client.$(selector);
+  await el.click();
+};
